use the role from the metadata
